tiktok_trend_analyzer/requirements.txt
flask>=2.0.1
anthropic>=0.7.0
python-dotenv>=0.19.0
SQLAlchemy>=2.0.25
requests>=2.26.0
Flask-SQLAlchemy>=3.1.1
selenium>=4.0.0
beautifulsoup4>=4.9.3
playwright>=1.41.0
APScheduler>=3.10.1 